requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
boto3>=1.28.0
httpx>=0.24.0
starlette>=0.27.0

